mcpclient/requirements.txt
mcp
openai
ollama
python-dotenv
